requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.0
attrs==23.2.0
black==24.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==5.0.3
django-dotenv==1.4.2
django-multifilefield==1.0
djangorestframework==3.15.0
frozenlist==1.4.1
idna==3.6
multidict==6.0.5
mypy-extensions==1.0.0
mysqlclient==2.2.4
packaging==24.0
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
psycopg2==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.31.0
sqlparse==0.4.4
twilio==9.0.2
tzdata==2024.1
urllib3==2.2.1
yarl==1.9.4
